requirements.txt
elevenlabs>=1.13.3
protobuf>=5.29.1
Requests>=2.32.3
google_cloud_texttospeech>=2.21.1
pydub>=0.25.1
audioop-lts; python_version>='3.13'